+++ noodles-csi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * csi/async/io/reader: Add async query reader (`csi::r#async::io::Query`).
+
 ## 0.51.0 - 2025-08-25
 
 ### Changed